Move ConstantTermNormalizer
* Move class ConstantTermNormalizer from SMTInterpol to Library-SMTLIB
* Extract method
* Add support for new representation of bitvector constants
  (ConstantTerm with BigInteger value)
* Add documentation

This class is currently not used by SMTInterpol. However, this class is
used by several projects of Ultimate, e.g., by the SMTSolverBridge.
Ultimate's projects should not depend on SMTInterpol or the SMTSolverBridge.
Hence Ultimate currently duplicates this class to make the code available
to its projects.
